Fixed debug print functions, added print unit tests

+++ src/mpack/mpack-reader.h
@@ -469,13 +469,13 @@ void mpack_discard(mpack_reader_t* reader);
  * Converts a blob of MessagePack to pseudo-JSON for debugging purposes
  * and pretty-prints it to the given file.
  */
-void mpack_print_file(const char* data, int len, FILE* file);
+void mpack_print_file(const char* data, size_t len, FILE* file);
 
 /**
  * Converts a blob of MessagePack to pseudo-JSON for debugging purposes
  * and pretty-prints it to stdout.
  */
-MPACK_INLINE void mpack_print(const char* data, int len) {
+MPACK_INLINE void mpack_print(const char* data, size_t len) {
     mpack_print_file(data, len, stdout);
 }
 #endif
